+++ lib/ghost-chef.rb
